+++ extensions/data-plane/data-plane-http/src/main/java/org/eclipse/edc/connector/dataplane/http/params/decorators/BaseCommonHttpParamsDecorator.java
@@ -1,5 +1,21 @@
-package org.eclipse.edc.connector.dataplane.http.params;
+/*
+ *  Copyright (c) 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
+ *
+ */
 
+package org.eclipse.edc.connector.dataplane.http.params.decorators;
+
+import org.eclipse.edc.connector.dataplane.http.params.HttpParamsDecorator;
+import org.eclipse.edc.connector.dataplane.http.params.HttpRequestParams;
 import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.security.Vault;
 import org.eclipse.edc.spi.types.TypeManager;
@@ -11,7 +27,7 @@ import java.util.Optional;
 
 import static java.lang.String.format;
 
-class BaseCommonHttpParamsDecorator implements HttpParamsDecorator {
+public class BaseCommonHttpParamsDecorator implements HttpParamsDecorator {
 
     private final Vault vault;
     private final TypeManager typeManager;
