+++ extensions/data-plane/data-plane-http/src/main/java/org/eclipse/edc/connector/dataplane/http/params/decorators/BaseSinkHttpParamsDecorator.java
@@ -1,11 +1,27 @@
-package org.eclipse.edc.connector.dataplane.http.params;
+/*
+ *  Copyright (c) 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
+ *
+ */
 
+package org.eclipse.edc.connector.dataplane.http.params.decorators;
+
+import org.eclipse.edc.connector.dataplane.http.params.HttpParamsDecorator;
+import org.eclipse.edc.connector.dataplane.http.params.HttpRequestParams;
 import org.eclipse.edc.spi.types.domain.HttpDataAddress;
 import org.eclipse.edc.spi.types.domain.transfer.DataFlowRequest;
 
 import java.util.Optional;
 
-class BaseSinkHttpParamsDecorator implements HttpParamsDecorator {
+public class BaseSinkHttpParamsDecorator implements HttpParamsDecorator {
     private static final String DEFAULT_METHOD = "POST";
 
     @Override
