+++ spi/data-plane/data-plane-spi/src/main/java/org/eclipse/edc/connector/dataplane/spi/pipeline/DataSource.java
@@ -72,8 +72,4 @@ public interface DataSource {
         }
     }
 
-    interface Builder {
-
-    }
-
 }
